Changed cmds to cogs. Rename test.py to utility.py
